#590 arrange Layers into packages and add explanation
