+++ layers/src/main/java/com/iluwatar/layers/dao/CakeLayerDao.java
@@ -21,33 +21,16 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.layers;
+package com.iluwatar.layers.dao;
 
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import com.iluwatar.layers.entity.CakeLayer;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
 
 /**
- * Date: 12/15/15 - 7:57 PM
- *
- * @author Jeroen Meulemeester
+ * CRUD repository for cake layers.
  */
-public class CakeBakingExceptionTest {
-
-  @Test
-  public void testConstructor() {
-    final CakeBakingException exception = new CakeBakingException();
-    assertNull(exception.getMessage());
-    assertNull(exception.getCause());
-  }
-
-  @Test
-  public void testConstructorWithMessage() {
-    final String expectedMessage = "message";
-    final CakeBakingException exception = new CakeBakingException(expectedMessage);
-    assertEquals(expectedMessage, exception.getMessage());
-    assertNull(exception.getCause());
-  }
+@Repository
+public interface CakeLayerDao extends CrudRepository<CakeLayer, Long> {
 
 }
