+++ layers/src/main/java/com/iluwatar/layers/dto/CakeLayerInfo.java
@@ -21,15 +21,39 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.layers;
+package com.iluwatar.layers.dto;
 
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
+import java.util.Optional;
 
 /**
- * CRUD repository for cakes.
+ * DTO for cake layers.
  */
-@Repository
-public interface CakeDao extends CrudRepository<Cake, Long> {
+public class CakeLayerInfo {
 
+  public final Optional<Long> id;
+  public final String name;
+  public final int calories;
+
+  /**
+   * Constructor.
+   */
+  public CakeLayerInfo(Long id, String name, int calories) {
+    this.id = Optional.of(id);
+    this.name = name;
+    this.calories = calories;
+  }
+
+  /**
+   * Constructor.
+   */
+  public CakeLayerInfo(String name, int calories) {
+    this.id = Optional.empty();
+    this.name = name;
+    this.calories = calories;
+  }
+
+  @Override
+  public String toString() {
+    return String.format("CakeLayerInfo id=%d name=%s calories=%d", id.orElse(-1L), name, calories);
+  }
 }
