+++ layers/src/main/java/com/iluwatar/layers/service/CakeBakingService.java
@@ -21,19 +21,46 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.layers;
+package com.iluwatar.layers.service;
+
+import com.iluwatar.layers.dto.CakeInfo;
+import com.iluwatar.layers.dto.CakeLayerInfo;
+import com.iluwatar.layers.dto.CakeToppingInfo;
+import com.iluwatar.layers.exception.CakeBakingException;
+import java.util.List;
 
 /**
- * Custom exception used in cake baking.
+ * Service for cake baking operations.
  */
-public class CakeBakingException extends Exception {
+public interface CakeBakingService {
 
-  private static final long serialVersionUID = 1L;
+  /**
+   * Bakes new cake according to parameters.
+   */
+  void bakeNewCake(CakeInfo cakeInfo) throws CakeBakingException;
 
-  public CakeBakingException() {
-  }
+  /**
+   * Get all cakes.
+   */
+  List<CakeInfo> getAllCakes();
 
-  public CakeBakingException(String message) {
-    super(message);
-  }
+  /**
+   * Store new cake topping.
+   */
+  void saveNewTopping(CakeToppingInfo toppingInfo);
+
+  /**
+   * Get available cake toppings.
+   */
+  List<CakeToppingInfo> getAvailableToppings();
+
+  /**
+   * Add new cake layer.
+   */
+  void saveNewLayer(CakeLayerInfo layerInfo);
+
+  /**
+   * Get available cake layers.
+   */
+  List<CakeLayerInfo> getAvailableLayers();
 }
