+++ layers/src/main/java/com/iluwatar/layers/view/CakeViewImpl.java
@@ -21,39 +21,26 @@
  * THE SOFTWARE.
  */
 
-package com.iluwatar.layers;
+package com.iluwatar.layers.view;
 
-import java.util.Optional;
+import com.iluwatar.layers.service.CakeBakingService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * DTO for cake layers.
+ * View implementation for displaying cakes.
  */
-public class CakeLayerInfo {
+public class CakeViewImpl implements View {
 
-  public final Optional<Long> id;
-  public final String name;
-  public final int calories;
+  private static final Logger LOGGER = LoggerFactory.getLogger(CakeViewImpl.class);
 
-  /**
-   * Constructor.
-   */
-  public CakeLayerInfo(Long id, String name, int calories) {
-    this.id = Optional.of(id);
-    this.name = name;
-    this.calories = calories;
+  private CakeBakingService cakeBakingService;
+
+  public CakeViewImpl(CakeBakingService cakeBakingService) {
+    this.cakeBakingService = cakeBakingService;
   }
 
-  /**
-   * Constructor.
-   */
-  public CakeLayerInfo(String name, int calories) {
-    this.id = Optional.empty();
-    this.name = name;
-    this.calories = calories;
-  }
-
-  @Override
-  public String toString() {
-    return String.format("CakeLayerInfo id=%d name=%s calories=%d", id.orElse(-1L), name, calories);
+  public void render() {
+    cakeBakingService.getAllCakes().forEach(cake -> LOGGER.info(cake.toString()));
   }
 }
